# learning/go/Section-19-Team-Practices-And-Tooling-Code-review-documentation-Makefile-DX-release-versioning-and-team-habits/Day-92/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Catalog API
  version: 1.2.0
  description: |
    Products and stock reservations.

    Errors always use the same body shape: {"error": "<code>", "message": "..."}.
    Switch on `error`; `message` is for humans and changes between releases.
servers:
  - url: http://localhost:8092
    description: local development
paths:
  /products:
    get:
      summary: List products
      operationId: listProducts
      responses:
        "200":
          description: Every product, ordered by SKU
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      summary: Add a product
      operationId: addProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewProduct"
      responses:
        "201":
          description: The product as stored, with updated_at set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          description: A product with this SKU already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  responses:
...
    InvalidRequest:
      description: The body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Product:
      type: object
      required: [sku, name, price_cent, stock, updated_at]
      properties:
        sku:
          type: string
          example: KB-01
        name:
          type: string
          example: Mechanical keyboard
        price_cent:
          type: integer
          format: int64
          description: Price in the smallest currency unit; never a float
          example: 12000
        stock:
          type: integer
          minimum: 0
        updated_at:
          type: string
          format: date-time
    NewProduct:
      type: object
      required: [sku, name, price_cent, stock]
      properties:
        sku:
          type: string
        name:
          type: string
        price_cent:
          type: integer
          format: int64
          minimum: 1
        stock:
          type: integer
          minimum: 0
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code
          enum: [not_found, conflict, invalid_request, internal]
        message:
          type: string
          description: Human-readable detail; may change between releases